projects/03_django_rest_framework/11_production_deployment/requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
dj-database-url==2.3.0
dj-email-url==1.0.6
dj-rest-auth==7.0.0
Django==5.1.1
django-allauth==65.2.0
django-cache-url==3.4.5
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
environs==11.2.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
marshmallow==3.23.1
packaging==24.1
psycopg2==2.9.10
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.7.0
